requirements.txt
bitarray>=0.8.1
numpy>=1.10.2
pandas>=0.17.1
pytz>=2016.0
pycapnp>=0.6.4
pytest>=3.0.0
pytest-flake8>=0.8.1
coveralls>=1.1
pytest-cov>=2.2.1
pytest-xdist
pytest-benchmark>=3.0.0
iso8601>=0.1.11
